refactor: replace anonymous conversations route with run-domain list endpoint
Remove the dedicated /run/anonymous/conversations endpoint and add a
proper /run/v1/conversations list endpoint that works for both
authenticated and anonymous users via the existing auth middleware.

- Delete anonymousConversations.ts and its custom JWE extraction
- Add conversations.ts in run routes using executionContext from middleware
- For anonymous users: uses SpiceDB lookupResources with DB fallback
- For authenticated users: lists conversations by tenant/project scope
- Remove isAnonymousRoute helper and middleware skips from createApp.ts
