+++ agents-api/src/domains/run/index.ts
@@ -1,9 +1,9 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
 import type { AppVariables } from '../../types';
 import agentRoutes from './routes/agents';
-import anonymousConversationRoutes from './routes/anonymousConversations';
 import chatRoutes from './routes/chat';
 import chatDataRoutes from './routes/chatDataStream';
+import conversationRoutes from './routes/conversations';
 import mcpRoutes from './routes/mcp';
 import webhookRoutes from './routes/webhooks';
 
@@ -12,8 +12,8 @@ export function createRunRoutes() {
 
   app.route('/v1/chat', chatRoutes);
   app.route('/api', chatDataRoutes);
+  app.route('/v1/conversations', conversationRoutes);
   app.route('/v1/mcp', mcpRoutes);
-  app.route('/anonymous', anonymousConversationRoutes);
   app.route('/agents', agentRoutes);
   app.route('/', webhookRoutes);
